requirements.txt
streamlit
numpy
opencv-python
joblib
scikit-learn
matplotlib
pillow
